# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API To-Do List",
        "description": "Documentação interativa da API para o gerenciador de tarefas.",
        "contact": {
            "email": "seu-email@exemplo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Servidor de Desenvolvimento Local"
        }
    ],
    "paths": {
...
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tarefas"
                ],
                "summary": "Lista todas as tarefas do usuário",
                "description": "Retorna uma lista de tarefas pertencentes ao usuário autenticado.",
                "operationId": "getTaskList",
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TaskResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tarefas"
                ],
                "summary": "Cria uma nova tarefa",
                "description": "Cria uma nova tarefa e a associa a categorias, se fornecido.",
                "operationId": "storeTask",
                "requestBody": {
                    "description": "Dados da nova tarefa",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Desenvolver nova feature"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Detalhes da implementação da feature X."
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31"
                                    },
                                    "priority": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "categories": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarefa criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "Tarefas"
                ],
                "summary": "Obtém informações de uma tarefa",
                "description": "Retorna os dados de uma tarefa específica.",
                "operationId": "getTaskById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da Tarefa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Recurso não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tarefas"
                ],
                "summary": "Exclui uma tarefa",
                "description": "Exclui uma tarefa existente.",
                "operationId": "deleteTask",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da Tarefa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tarefa excluída com sucesso"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Recurso não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Tarefas"
                ],
                "summary": "Atualiza uma tarefa existente (parcial)",
                "description": "Atualiza os dados de uma tarefa existente. Apenas os campos fornecidos serão atualizados.",
                "operationId": "updateTask",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da Tarefa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Novos dados da tarefa",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Revisar documentação"
                                    },
                                    "completed": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "priority": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "categories": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarefa atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Recurso não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "CategoryResource": {
                "title": "Recurso de Categoria",
                "properties": {
                    "id": {
                        "description": "ID da categoria",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nome da categoria",
                        "type": "string",
                        "example": "Pessoal"
                    },
                    "color": {
                        "description": "Cor da categoria em formato hexadecimal",
                        "type": "string",
                        "format": "hex",
                        "example": "#D0021B"
                    }
                },
                "type": "object"
            },
            "TaskResource": {
                "title": "Recurso de Tarefa",
                "properties": {
                    "id": {
                        "description": "ID da tarefa",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Título da tarefa",
                        "type": "string",
                        "example": "Minha Primeira Tarefa"
                    },
                    "description": {
                        "description": "Descrição detalhada da tarefa",
                        "type": "string",
                        "example": "Fazer algo importante.",
                        "nullable": true
                    },
                    "completed": {
                        "description": "Indica se a tarefa foi concluída",
                        "type": "boolean",
                        "example": false
                    },
                    "due_date": {
                        "description": "Data de vencimento",
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-31",
                        "nullable": true
                    },
                    "priority": {
                        "description": "Nível de prioridade (1-5)",
                        "type": "integer",
                        "example": 4,
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time"
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CategoryResource"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}